Move publish command into own file
